+++ rusoto/services/docdb/src/lib.rs
@@ -1,0 +1,30 @@
+
+// =================================================================
+//
+//                           * WARNING *
+//
+//                    This file is generated!
+//
+//  Changes made to this file will be overwritten. If changes are
+//  required to the generated code, the service_crategen project
+//  must be updated to generate the changes.
+//
+// =================================================================
+
+#![doc(html_logo_url = "https://raw.githubusercontent.com/rusoto/rusoto/master/assets/logo-square.png")]
+//! <p>Amazon DocumentDB API documentation</p>
+//!
+//! If you're using the service, you're probably looking for [DocdbClient](struct.DocdbClient.html) and [Docdb](trait.Docdb.html).
+
+extern crate bytes;
+extern crate futures;
+extern crate rusoto_core;
+extern crate serde_urlencoded;
+extern crate xml;
+
+mod generated;
+mod custom;
+
+pub use generated::*;
+pub use custom::*;
+