+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Fix de/serialization of DynamoDB binary set attribute values
 - Change type for blob values from `Vec<u8>` to `Bytes`
+- Add DocDB service
 - Add Kafka service
 
 ## [0.38.0] - 2019-04-17